Home hero animation reorganization + bugfixing
