+++ app/layout/home/hero/hero-animation/index.ts
@@ -1,0 +1,2 @@
+export { BACKGROUND_IMAGES, MAIN_IMAGES } from './constants';
+export { default } from './component';
